File clean up to only have two files (deleted the rest) of proof of concept of the flow, the pipeline & the complex flows.
